# openapi.json
{
  "openapi" : "3.0.1",
  "info" : {
    "title" : "${springdoc.title}",
    "description" : "Documentation ${spring.application.name} v1.0",
    "version" : "${springdoc.version}"
  },
  "servers" : [ {
    "url" : "http://172.20.10.9:8080",
    "description" : "Generated server url"
  } ],
  "paths" : {
    "/api/v1/customers/{id}" : {
      "get" : {
        "tags" : [ "customer-controller" ],
        "summary" : "Find customer by ID",
        "description" : "Returns a single customer",
        "operationId" : "getCustomerById",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/APIResponseCustomerDTO"
                }
              }
            }
          }
        }
      },
      "put" : {
        "tags" : [ "customer-controller" ],
        "summary" : "Update an existing customer",
        "operationId" : "updateCustomer",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CustomerRequestDTO"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/APIResponseCustomerDTO"
                }
              }
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "customer-controller" ],
        "summary" : "Delete a customer by ID",
        "operationId" : "deleteCustomer",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/APIResponseString"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/customers" : {
      "post" : {
        "tags" : [ "customer-controller" ],
        "summary" : "Add a new customer",
        "operationId" : "createCustomer",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CustomerRequestDTO"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/APIResponseCustomerDTO"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/customers/search" : {
      "post" : {
        "tags" : [ "customer-controller" ],
        "summary" : "Search customers with pagination",
        "description" : "Returns a paginated list of customers based on the search criteria",
        "operationId" : "getAllCustomersUsingPagination",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CustomerSearchCriteriaDTO"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/APIResponsePageCustomerDTO"
                }
              }
            }
          }
        }
      }
    }
  },
  "components" : {
    "schemas" : {
      "CustomerRequestDTO" : {
        "required" : [ "dateOfBirth", "email", "firstName", "lastName" ],
        "type" : "object",
        "properties" : {
          "firstName" : {
            "maxLength" : 2147483647,
            "minLength" : 2,
            "type" : "string"
          },
          "lastName" : {
            "maxLength" : 2147483647,
            "minLength" : 2,
            "type" : "string"
          },
          "email" : {
            "type" : "string"
          },
          "phoneNumber" : {
            "maxLength" : 10,
            "minLength" : 10,
            "type" : "string"
          },
          "dateOfBirth" : {
            "type" : "string",
            "format" : "date"
          }
        }
      },
      "APIResponseCustomerDTO" : {
        "type" : "object",
        "properties" : {
          "status" : {
            "type" : "string"
          },
          "errors" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/ErrorDTO"
            }
          },
          "results" : {
            "$ref" : "#/components/schemas/CustomerDTO"
          }
        }
      },
      "CustomerDTO" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "firstName" : {
            "type" : "string"
          },
          "lastName" : {
            "type" : "string"
          },
          "email" : {
            "type" : "string"
          },
          "phoneNumber" : {
            "type" : "string"
          },
          "dateOfBirth" : {
            "type" : "string",
            "format" : "date"
          }
        }
      },
      "ErrorDTO" : {
        "type" : "object",
        "properties" : {
          "field" : {
            "type" : "string"
          },
          "errorMessage" : {
            "type" : "string"
          }
        }
      },
      "CustomerSearchCriteriaDTO" : {
        "required" : [ "page", "size" ],
        "type" : "object",
        "properties" : {
          "firstName" : {
            "type" : "string"
          },
          "lastName" : {
            "type" : "string"
          },
          "email" : {
            "type" : "string"
          },
          "phoneNumber" : {
            "type" : "string"
          },
          "dateOfBirth" : {
            "type" : "string",
            "format" : "date"
          },
          "page" : {
            "type" : "integer",
            "format" : "int32"
          },
          "size" : {
            "type" : "integer",
            "format" : "int32",
            "example" : 10
          },
          "sortList" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/SortItem"
            }
          }
        }
      },
      "SortItem" : {
        "type" : "object",
        "properties" : {
          "field" : {
            "type" : "string",
            "example" : "id"
          },
          "direction" : {
            "type" : "string",
            "enum" : [ "ASC", "DESC" ]
          }
        }
      },
      "APIResponsePageCustomerDTO" : {
        "type" : "object",
        "properties" : {
          "status" : {
            "type" : "string"
          },
          "errors" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/ErrorDTO"
            }
          },
          "results" : {
            "$ref" : "#/components/schemas/PageCustomerDTO"
          }
        }
      },
      "PageCustomerDTO" : {
        "type" : "object",
        "properties" : {
          "totalPages" : {
            "type" : "integer",
            "format" : "int32"
          },
          "totalElements" : {
            "type" : "integer",
            "format" : "int64"
          },
          "first" : {
            "type" : "boolean"
          },
          "last" : {
            "type" : "boolean"
          },
          "size" : {
            "type" : "integer",
            "format" : "int32"
          },
          "content" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/CustomerDTO"
            }
          },
          "number" : {
            "type" : "integer",
            "format" : "int32"
          },
          "sort" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/SortObject"
            }
          },
          "pageable" : {
            "$ref" : "#/components/schemas/PageableObject"
          },
          "numberOfElements" : {
            "type" : "integer",
            "format" : "int32"
          },
          "empty" : {
            "type" : "boolean"
          }
        }
      },
      "PageableObject" : {
        "type" : "object",
        "properties" : {
          "offset" : {
            "type" : "integer",
            "format" : "int64"
          },
          "sort" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/SortObject"
            }
          },
          "paged" : {
            "type" : "boolean"
          },
          "pageNumber" : {
            "type" : "integer",
            "format" : "int32"
          },
          "pageSize" : {
            "type" : "integer",
            "format" : "int32"
          },
          "unpaged" : {
            "type" : "boolean"
          }
        }
      },
      "SortObject" : {
        "type" : "object",
        "properties" : {
          "direction" : {
            "type" : "string"
          },
          "nullHandling" : {
            "type" : "string"
          },
          "ascending" : {
            "type" : "boolean"
          },
          "property" : {
            "type" : "string"
          },
          "ignoreCase" : {
            "type" : "boolean"
          }
        }
      },
      "APIResponseString" : {
        "type" : "object",
        "properties" : {
          "status" : {
            "type" : "string"
          },
          "errors" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/ErrorDTO"
            }
          },
          "results" : {
            "type" : "string"
          }
        }
      }
    }
  }
}